# proto/gen/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/v1/workspace/ai/test-agent:
        post:
            tags:
                - WorkspaceService
            description: Tests the connection to a specific AI agent.
            operationId: WorkspaceService_TestAIAgent
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/TestAIAgentRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TestAIAgentResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/workspace/profile:
        get:
            tags:
                - WorkspaceService
            description: Gets the workspace profile.
            operationId: WorkspaceService_GetWorkspaceProfile
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/WorkspaceProfile'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
components:
    schemas:
...
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
...
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
...
        TestAIAgentRequest:
            required:
                - apiKey
                - endpoint
                - agentId
            type: object
            properties:
                apiKey:
                    type: string
                    description: api_key is the Bailian API Key to test.
                endpoint:
                    type: string
                    description: endpoint is the API endpoint (e.g., "dashscope.aliyuncs.com").
                agentId:
                    type: string
                    description: agent_id is the specific agent ID to test.
            description: Request message for TestAIAgent method.
        TestAIAgentResponse:
            type: object
            properties:
                success:
                    type: boolean
                    description: success indicates whether the test passed.
                message:
                    type: string
                    description: message contains the test result message.
                latencyMs:
                    type: integer
                    description: latency_ms is the test latency in milliseconds (optional).
                    format: int32
            description: Response message for TestAIAgent method.
...
        WorkspaceProfile:
            type: object
            properties:
                owner:
                    type: string
                    description: |-
                        The name of instance owner.
                         Format: users/{user}
                version:
                    type: string
                    description: Version is the current version of instance.
                mode:
                    type: string
                    description: Mode is the instance mode (e.g. "prod", "dev" or "demo").
                instanceUrl:
                    type: string
                    description: Instance URL is the URL of the instance.
            description: Workspace profile message containing basic workspace information.
...